:memo: Task Management | Dashboard improvement, responsive
