+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "timetracking",
+  "name": "time-tracker",
   "author": "Quentin Aslan (quentin.aslan@outlook.com)",
   "private": true,
   "version": "0.1.0",
